+++ src/settings.cpp
@@ -56,7 +56,7 @@ string REQUEST_DATA_FILE = "requests.csv";
 string LEG_REQUEST_DATA_FILE = "leg_requests.csv";
 string RESULTS_DIRECTORY = "results";
 int RTV_TIMELIMIT = 0;
-int GRB_TIME_LIMIT = 90;
+double GRB_TIME_LIMIT = 90.0;
 string TIMEFILE = "times.csv";
 string VEHICLE_DATA_FILE = "vehicles.csv";
 int VEHICLE_LIMIT = 1000; // 0;
@@ -170,7 +170,7 @@ void initialize(int argc, char** argv)
         else if (key == "RTV_TIMELIMIT")
             RTV_TIMELIMIT = stoi(value);
         else if (key == "GRB_TIME_LIMIT")
-            GRB_TIME_LIMIT = stoi(value);
+            GRB_TIME_LIMIT = stod(value);
         else if (key == "DWELL_PICKUP")
             DWELL_PICKUP = stoi(value);
         else if (key == "DWELL_ALIGHT")
